+++ ndhistogram/src/histogram/hashhistogram.rs
@@ -2,7 +2,7 @@ use std::{
     collections::HashMap,
     collections::HashSet,
     fmt::Display,
-    ops::{Add, Div, Mul, Sub},
+    ops::{Add, AddAssign, Div, DivAssign, Mul, MulAssign, Sub, SubAssign},
 };
 
 use serde::{Deserialize, Serialize};
@@ -167,8 +167,8 @@ where
     }
 }
 
-macro_rules! impl_binary_op {
-    ($Trait:tt, $method:tt, $mathsymbol:tt) => {
+macro_rules! impl_binary_op_with_immutable_borrow {
+    ($Trait:tt, $method:tt, $mathsymbol:tt, $testresult:tt) => {
         impl<A: Axis + PartialEq + Clone, V> $Trait<&HashHistogram<A, V>> for &HashHistogram<A, V>
         where
             HashHistogram<A, V>: Histogram<A, V>,
@@ -177,6 +177,22 @@ macro_rules! impl_binary_op {
         {
             type Output = Result<HashHistogram<A, V>, BinaryOperationError>;
 
+            /// Combine the right-hand histogram with the left-hand histogram,
+            /// returning a copy, and leaving the original histograms intact.
+            ///
+            /// If the input histograms have incompatible axes, this operation
+            /// will return a [BinaryOperationError].
+            ///
+            /// # Examples
+            ///
+            /// ```rust
+            /// use ndhistogram::{Histogram, sparsehistogram, axis::Uniform};
+            /// let mut hist1 = sparsehistogram!(Uniform::<f64>::new(10, -5.0, 5.0));
+            /// let mut hist2 = sparsehistogram!(Uniform::<f64>::new(10, -5.0, 5.0));
+            /// hist1.fill_with(&0.0, 2.0);
+            /// hist2.fill(&0.0);
+            #[doc=concat!("let combined_hist = (&hist1 ", stringify!($mathsymbol), " &hist2).expect(\"Axes are compatible\");")]
+            #[doc=concat!("assert_eq!(combined_hist.value(&0.0).unwrap(), &", stringify!($testresult), ");")]
             fn $method(self, rhs: &HashHistogram<A, V>) -> Self::Output {
                 if self.axes() != rhs.axes() {
                     return Err(BinaryOperationError);
@@ -202,7 +218,100 @@ macro_rules! impl_binary_op {
     };
 }
 
-impl_binary_op! {Add, add, +}
-impl_binary_op! {Sub, sub, -}
-impl_binary_op! {Mul, mul, *}
-impl_binary_op! {Div, div, /}
+impl_binary_op_with_immutable_borrow! {Add, add, +, 3.0}
+impl_binary_op_with_immutable_borrow! {Sub, sub, -, 1.0}
+impl_binary_op_with_immutable_borrow! {Mul, mul, *, 2.0}
+impl_binary_op_with_immutable_borrow! {Div, div, /, 2.0}
+
+macro_rules! impl_binary_op_with_owned {
+    ($Trait:tt, $method:tt, $ValueAssignTrait:tt, $mathsymbol:tt, $assignmathsymbol:tt, $testresult:tt) => {
+        impl<A: Axis + PartialEq + Clone, V> $Trait<&HashHistogram<A, V>> for HashHistogram<A, V>
+        where
+            HashHistogram<A, V>: Histogram<A, V>,
+            V: Clone + Default,
+            for<'a> V: $ValueAssignTrait<&'a V>,
+        {
+            type Output = Result<HashHistogram<A, V>, BinaryOperationError>;
+
+            /// Combine the right-hand histogram with the left-hand histogram,
+            /// consuming the left-hand histogram and returning a new value.
+            /// As this avoids making copies of the histograms, this is the
+            /// recommended method to merge histograms.
+            ///
+            /// If the input histograms have incompatible axes, this operation
+            /// will return a [BinaryOperationError].
+            ///
+            /// # Examples
+            ///
+            /// ```rust
+            /// use ndhistogram::{Histogram, sparsehistogram, axis::Uniform};
+            /// let mut hist1 = sparsehistogram!(Uniform::<f64>::new(10, -5.0, 5.0));
+            /// let mut hist2 = sparsehistogram!(Uniform::<f64>::new(10, -5.0, 5.0));
+            /// hist1.fill_with(&0.0, 2.0);
+            /// hist2.fill(&0.0);
+            #[doc=concat!("let combined_hist = (hist1 ", stringify!($mathsymbol), " &hist2).expect(\"Axes are compatible\");")]
+            #[doc=concat!("assert_eq!(combined_hist.value(&0.0).unwrap(), &", stringify!($testresult), ");")]
+            fn $method(mut self, rhs: &HashHistogram<A, V>) -> Self::Output {
+                if self.axes() != rhs.axes() {
+                    return Err(BinaryOperationError);
+                }
+                for (index, rhs_value) in rhs.values.iter() {
+                    let lhs_value = self.values.entry(*index).or_default();
+                    *lhs_value $assignmathsymbol rhs_value
+                }
+                Ok(self)
+            }
+        }
+    };
+}
+
+impl_binary_op_with_owned! {Add, add, AddAssign, +, +=, 3.0}
+impl_binary_op_with_owned! {Sub, sub, SubAssign, -, -=, 1.0}
+impl_binary_op_with_owned! {Mul, mul, MulAssign, *, *=, 2.0}
+impl_binary_op_with_owned! {Div, div, DivAssign, /, /=, 2.0}
+
+macro_rules! impl_binary_op_assign {
+    ($Trait:tt, $method:tt, $ValueAssignTrait:tt, $mathsymbol:tt, $testresult:tt) => {
+        impl<A: Axis + PartialEq, V> $Trait<&HashHistogram<A, V>> for HashHistogram<A, V>
+        where
+            HashHistogram<A, V>: Histogram<A, V>,
+            V: Default,
+            for<'a> V: $ValueAssignTrait<&'a V>,
+        {
+
+           /// Combine the right-hand histogram with the left-hand histogram,
+            /// mutating the left-hand histogram.
+            ///
+            /// # Panics
+            ///
+            /// Panics if the histograms have incompatible axes.
+            /// To handle this failure mode at runtime, use the non-assign
+            /// version of this operation, which returns an Result.
+            ///
+            /// # Examples
+            ///
+            /// ```rust
+            /// use ndhistogram::{Histogram, sparsehistogram, axis::Uniform};
+            /// let mut hist1 = sparsehistogram!(Uniform::<f64>::new(10, -5.0, 5.0));
+            /// let mut hist2 = sparsehistogram!(Uniform::<f64>::new(10, -5.0, 5.0));
+            /// hist1.fill_with(&0.0, 2.0);
+            /// hist2.fill(&0.0);
+            #[doc=concat!("hist1 ", stringify!($mathsymbol), " &hist2;")]
+            #[doc=concat!("assert_eq!(hist1.value(&0.0).unwrap(), &", stringify!($testresult), ");")]
+            fn $method(&mut self, rhs: &HashHistogram<A, V>) {
+                if self.axes() != rhs.axes() {
+                    panic!("Cannot combine HashHistograms with incompatible axes.");
+                }
+                for (index, rhs_value) in rhs.values.iter() {
+                    let lhs_value = self.values.entry(*index).or_default();
+                    *lhs_value $mathsymbol rhs_value
+                }
+            }
+        }
+    };
+}
+
+impl_binary_op_assign! {AddAssign, add_assign, AddAssign, +=, 3.0}
+impl_binary_op_assign! {SubAssign, sub_assign, SubAssign, -=, 1.0}
+impl_binary_op_assign! {MulAssign, mul_assign, MulAssign, *=, 2.0}
+impl_binary_op_assign! {DivAssign, div_assign, DivAssign, /=, 2.0}